+++ src/main/java/ru/netology/servlet/MainServlet.java
@@ -3,7 +3,7 @@ package ru.netology.servlet;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import ru.netology.controller.PostController;
 import ru.netology.exception.NotFoundException;
-import ru.netology.repository.PostRepository;
+import ru.netology.repository.PostRepositoryImpl;
 import ru.netology.service.PostService;
 
 import javax.servlet.http.HttpServlet;
@@ -26,7 +26,7 @@ public class MainServlet extends HttpServlet {
      */
     @Override
     public void init() {
-        final var repository = (PostRepository) context.getBean("postRepository");
+        final var repository = (PostRepositoryImpl) context.getBean("postRepositoryImpl");
         final var service = (PostService) context.getBean("postService");
         controller = (PostController) context.getBean("postController");
     }
